Updated for better chance of downloading, fixed problem with titles and images, added usage, parameterized

+++ gather.sh
@@ -13,28 +13,45 @@
 # Doesn't re-download stuff that's already done. You are responsible for cleaning up the directory
 # periodically. 
 # 
-cd reddit_image_gatherer
-curl -so /tmp/$$_tmp1.txt 'http://www.reddit.com/r/EarthPorn+SpacePorn+AutumnPorn+WinterPorn+WaterPorn+ArchitecturePorn/search.rss?q=1920x1200&restrict_sr=on&limit=100' 
-xmllint --pretty 1 /tmp/$$_tmp1.txt >/tmp/$$_tmp2.txt
-grep '\<description\>' /tmp/$$_tmp2.txt | awk -F= '{ for ( i=1; i < NF; i++) { printf("%s\n",$i); } }' | egrep '(" title|link])' | awk -F\" '{ print $2; }'  >/tmp/$$_tmp3.txt
-cd images
+if [ $# -ne 2 ]; then
+    echo "Usage: gather.sh images_dirname URL"
+    echo "  Make sure you quote things properly, e.g."
+    echo "  gather.sh '/Users/my name/my images' 'http://www.reddit.com/r/EarthPorn+SpacePorn+AutumnPorn+WinterPorn+WaterPorn+ArchitecturePorn/search.rss?restrict_sr=on&limit=100'"
+    exit
+fi
+echo $1 $2
+cd "$1"
+curl -so /tmp/$$_tmp1.txt $2
+/usr/bin/xmllint --pretty 1 /tmp/$$_tmp1.txt | egrep '(\<title\>|\<description\>)' |grep -v 'media:title' |grep -v "multi: search">/tmp/$$_tmp3.txt
 while  read line 
 do
-	#/bin/echo -n "#TITLE#"$line"##"
-	title=$line
+    #echo "######################"
+    if [ "$line" = "<description/>" ]; then
+	read line
+    fi
+	#/bin/echo "#TITLE#"$line"##"
+	title=`echo $line | sed 's/\<title\>//g' | sed 's/\<\/title\>//g'`
 	read line
 	#echo "#IMAGE#"$line"##"
-	image=$line
-	echo $image | grep http | grep -v flickr |egrep '(png|jpg|jpeg)' >&/dev/null
+	image=`echo $line | awk -F= '{ for ( i=1; i < NF; i++) { printf("%s\n",$i); } }' | egrep 'link]' | awk -F\" '{ print $2; }'`
+	if [ $? -eq 0 ]; then
+  	    echo $image | egrep '(png|jpg|jpeg)' >&/dev/null
+	    if [ $? -ne 0 ]; then
+	        # sometimes imgur links are directly to the image
+		image=${image}.jpg
+	    fi
+	fi
+	# flickr galleries don't work, static flickr links do work
+	echo $image | grep http | grep -v www.flickr.com |egrep '(png|jpg|jpeg)' >&/dev/null
 	if [ $? -eq 0 ]; then
 		filename=`echo $image | awk -F/ '{ print $NF; }'`
-		/bin/echo -n "Download and process [$image:$title]... "
+		/bin/echo -n "Download and process [$image:$title:$filename]... "
 		if [ -f titled_$filename ]; then
 			echo already done.
 		else
 			curl -so $filename $image
 			if [ -f $filename ]; then
-				php ../imgedit.php $filename "$title" 
+				php `dirname $0`/imgedit.php $filename "$title" 
 				rm $filename
 				echo done.
 			else
